# docs/swagger.json
{
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "T-Loyal API",
        "title": "Crazy Peppers T-Loyal API",
        "contact": {},
        "version": "1.0"
    },
    "host": "prod-team-29-4254c2ee.REDACTED",
    "basePath": "/api/v1",
    "paths": {
        "/clients/auth/sign-in": {
            "post": {
                "description": "sign-in in client with given data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clients"
                ],
                "summary": "Sign-in for client",
                "parameters": [
                    {
                        "description": "Credentials to use",
                        "name": "client",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.signInRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.clientSignResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    }
                }
            }
        },
        "/clients/auth/sign-up": {
            "post": {
                "description": "create client with given data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clients"
                ],
                "summary": "Create client",
                "parameters": [
                    {
                        "description": "Credentials to use",
                        "name": "client",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Client"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.clientSignResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/moderation": {
            "get": {
                "description": "Get list for moderation",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Moderation"
                ],
                "summary": "Get not approved endpoint",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Promo"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    }
                }
            },
            "post": {
                "description": "Change moderation status for promo",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Moderation"
                ],
                "summary": "Moderate promo",
                "parameters": [
                    {
                        "description": "Verdict of moderator",
                        "name": "Verdict",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ModerationBody"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    }
                }
            }
        },
...
        "/partners/auth/sign-in": {
            "post": {
                "description": "sign-in in partners with given data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Partners"
                ],
                "summary": "Sign-in for partners",
                "parameters": [
                    {
                        "description": "Credentials to use",
                        "name": "partner",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.signInRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.partnerSignResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    }
                }
            }
        },
        "/partners/auth/sign-up": {
            "post": {
                "description": "sign-up partner with given data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Partners"
                ],
                "summary": "Sign-up partner",
                "parameters": [
                    {
                        "description": "Credentials to use",
                        "name": "partner",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Partner"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.partnerSignResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    }
                }
            }
        },
...
        "/promos/activate": {
            "post": {
                "description": "Activate promo for client",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PromosForPartners"
                ],
                "summary": "Promo endpoint",
                "parameters": [
                    {
                        "description": "Activation Request",
                        "name": "promo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ActivationRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization header must be set for access",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Promo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.clientSignResponse": {
            "type": "object",
            "required": [
                "name",
                "phone_number"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 3
                },
                "password": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "handlers.partnerSignResponse": {
            "type": "object",
            "required": [
                "description",
                "image_id",
                "phone_number",
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 500,
                    "minLength": 3
                },
                "id": {
                    "type": "string"
                },
                "image_id": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 3
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "handlers.signInRequest": {
            "type": "object",
            "required": [
                "password",
                "phone_number"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
        "models.ActivationRequest": {
            "type": "object",
            "properties": {
                "client_id": {
                    "type": "string"
                },
                "promo_id": {
                    "type": "string"
                }
            }
        },
        "models.Client": {
            "type": "object",
            "required": [
                "name",
                "phone_number"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 3
                },
                "password": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
        "models.Condition": {
            "type": "object",
            "properties": {
                "buy": {
                    "type": "integer"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "COMMON",
                        "UNIQUE",
                        "BUNDLE"
                    ]
                }
            }
        },
...
        "models.ModerationBody": {
            "type": "object",
            "properties": {
                "partner_id": {
                    "type": "string"
                },
                "promo_id": {
                    "type": "string"
                },
                "verdict": {
                    "type": "boolean"
                }
            }
        },
        "models.Partner": {
            "type": "object",
            "required": [
                "description",
                "image_id",
                "phone_number",
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 500,
                    "minLength": 3
                },
                "id": {
                    "type": "string"
                },
                "image_id": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 3
                }
            }
        },
        "models.Promo": {
            "type": "object",
            "required": [
                "condition",
                "image_id"
            ],
            "properties": {
                "condition": {
                    "$ref": "#/definitions/models.Condition"
                },
                "condition_string": {
                    "description": "✅ ИСПРАВЛЕНО ЗДЕСЬ!",
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "image_id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "is_approved": {
                    "type": "boolean"
                },
                "partner_id": {
                    "type": "string"
                },
                "promo_id": {
                    "type": "string"
                },
                "time_limit_end": {
                    "type": "integer"
                },
                "time_limit_start": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "usage_limit": {
                    "type": "integer"
                }
            }
        },
...
        "utils.Err": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}